// indexer/src/types/Button.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */
import type {
  BaseContract,
  BigNumberish,
  BytesLike,
  FunctionFragment,
  Result,
  Interface,
  EventFragment,
  AddressLike,
  ContractRunner,
  ContractMethod,
  Listener,
} from "ethers";
import type {
  TypedContractEvent,
  TypedDeferredTopicFilter,
  TypedEventLog,
  TypedLogDescription,
  TypedListener,
  TypedContractMethod,
} from "./common";

export interface ButtonInterface extends Interface {
  getFunction(
    nameOrSignature: "lastPressTime" | "pressButton"
  ): FunctionFragment;

  getEvent(nameOrSignatureOrTopic: "ButtonPressed"): EventFragment;

  encodeFunctionData(
    functionFragment: "lastPressTime",
    values: [AddressLike]
  ): string;
  encodeFunctionData(
    functionFragment: "pressButton",
    values?: undefined
  ): string;

  decodeFunctionResult(
    functionFragment: "lastPressTime",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "pressButton",
    data: BytesLike
  ): Result;
}

export namespace ButtonPressedEvent {
  export type InputTuple = [user: AddressLike, timestamp: BigNumberish];
  export type OutputTuple = [user: string, timestamp: bigint];
  export interface OutputObject {
    user: string;
    timestamp: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface Button extends BaseContract {
  connect(runner?: ContractRunner | null): Button;
  waitForDeployment(): Promise<this>;

  interface: ButtonInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  lastPressTime: TypedContractMethod<[arg0: AddressLike], [bigint], "view">;

  pressButton: TypedContractMethod<[], [void], "nonpayable">;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "lastPressTime"
  ): TypedContractMethod<[arg0: AddressLike], [bigint], "view">;
  getFunction(
    nameOrSignature: "pressButton"
  ): TypedContractMethod<[], [void], "nonpayable">;

  getEvent(
    key: "ButtonPressed"
  ): TypedContractEvent<
    ButtonPressedEvent.InputTuple,
    ButtonPressedEvent.OutputTuple,
    ButtonPressedEvent.OutputObject
  >;

  filters: {
    "ButtonPressed(address,uint256)": TypedContractEvent<
      ButtonPressedEvent.InputTuple,
      ButtonPressedEvent.OutputTuple,
      ButtonPressedEvent.OutputObject
    >;
    ButtonPressed: TypedContractEvent<
      ButtonPressedEvent.InputTuple,
      ButtonPressedEvent.OutputTuple,
      ButtonPressedEvent.OutputObject
    >;
  };
}